lectures/lecture-4/auth/requirements.txt
fastapi[all]
pydantic
sqlalchemy
alembic
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
